requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
flask-cors==4.0.0
flask-restx==1.3.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2 